Refactor: move base class and reuse in switch
