+++ custom_components/chargeamps/sensor.py
@@ -5,7 +5,8 @@ import logging
 from homeassistant.const import DEVICE_CLASS_POWER, ENERGY_KILO_WATT_HOUR, POWER_WATT
 from homeassistant.helpers.entity import Entity
 
-from .const import DOMAIN, DOMAIN_DATA, ICON, ICON_ENERGY
+from . import ChargeampsEntity
+from .const import DOMAIN, DOMAIN_DATA
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -46,60 +47,11 @@ async def async_setup_platform(
     async_add_entities(sensors, True)
 
 
-class ChargeampsEntity(Entity):
-    """Chargeamps Entity class."""
-
-    def __init__(self, hass, name, charge_point_id, connector_id):
-        self.hass = hass
-        self.charge_point_id = charge_point_id
-        self.connector_id = connector_id
-        self.handler = self.hass.data[DOMAIN_DATA]["handler"]
-        self._name = name
-        self._state = None
-        self._attributes = {}
-        self._interviewed = False
-
-    async def interview(self):
-        chargepoint_info = self.handler.get_chargepoint_info(self.charge_point_id)
-        connector_info = self.handler.get_connector_info(
-            self.charge_point_id, self.connector_id
-        )
-        self._attributes["chargepoint_type"] = chargepoint_info.type
-        self._attributes["connector_type"] = connector_info.type
-        self._interviewed = True
-
-    @property
-    def name(self):
-        """Return the name of the sensor."""
-        return self._name
-
-    @property
-    def state(self):
-        """Return the state of the sensor."""
-        return self._state
-
-    @property
-    def device_state_attributes(self):
-        """Return the state attributes of the sensor."""
-        return self._attributes
-
-    @property
-    def unique_id(self):
-        """Return a unique ID to use for this sensor."""
-        return f"{DOMAIN}_{self.charge_point_id}_{self.connector_id}"
-
-
 class ChargeampsSensor(ChargeampsEntity):
     """Chargeamps Sensor class."""
 
     def __init__(self, hass, name, charge_point_id, connector_id):
         super().__init__(hass, name, charge_point_id, connector_id)
-        self._icon = ICON
-
-    @property
-    def icon(self):
-        """Icon to use in the frontend, if any."""
-        return self._icon
 
     async def async_update(self):
         """Update the sensor."""
@@ -132,11 +84,6 @@ class ChargeampsTotalEnergy(ChargeampsEntity):
 
     def __init__(self, hass, name, charge_point_id):
         super().__init__(hass, name, charge_point_id, "total_energy")
-
-    @property
-    def icon(self):
-        """Icon to use in the frontend, if any."""
-        return ICON_ENERGY
 
     async def async_update(self):
         """Update the sensor."""
